decided not to implement Integral for Arith
While the code is ok, the problem is that Integral implies Eq, and rather a
lot of function on Integrals end up checking if they == 0, etc. And, the Eq
instance for Arith doesn't compare the value it would evalulate to, but
only the structure. So, this Integral instance could allow writing buggy
code. Better for ADiv to be used instead of quot.