Improvements to build with base-4.12 (MonadFail)
Problem was this line:

	tmp@(VarTerm internal) <- newVarUnsafe shortname :: Script (Term Var ())

Despite newVarUnsafe being guarenteed to return a Term Var, so the
constructor must be VarTerm and never StaticTerm, ghc flags this as needing
MonadFail. I don't entirely understand why.

Considered adding MonadFail to Script, and it might make sense for other
reasons, but it's surely not needed for here. So, refactored code to
use UntypedVar. That also avoided some unwrapping/wrapping it was doing.